allow dropdownAutoComplete to manage state and have whatever control it
